style: 로그인 배경화면 이미지,  color 셋팅
