+++ app/private/_layout.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { View } from 'react-native';
+
+function PrivateLayout() {
+  return <View>레이아웃</View>;
+}
+
+export default PrivateLayout;
