+++ styles/color.ts
@@ -8,3 +8,14 @@ export const gray = {
   90: '#252525',
   100: '#121212',
 };
+
+export const main = {
+  10: '#F1E7FF',
+  20: '#D9C5FE',
+  30: '#BF9EFF',
+  50: '#A373FF',
+  60: '#8A50FF',
+  80: '#7028FF',
+  90: '#4814AF',
+  100: '#320C7D',
+};
